# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise HRMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public-holidays:
    get:
      tags: [leave]
      summary: List public holidays
      operationId: listPublicHolidays
      parameters:
        - name: year
          in: query
          schema: { type: number }
        - name: applicableTo
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PublicHoliday" }
    post:
      tags: [leave]
      summary: Create public holiday
      operationId: createPublicHoliday
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicHolidayInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicHoliday" }
  /public-holidays/{id}:
    patch:
      tags: [leave]
      summary: Update public holiday
      operationId: updatePublicHoliday
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicHolidayInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicHoliday" }
    delete:
      tags: [leave]
      summary: Delete public holiday
      operationId: deletePublicHoliday
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: No Content

  # ── Salary Grades ─────────────────────────────────────────────────────────
  /salary-grades:
    get:
      tags: [payroll]
      summary: List salary grades
      operationId: listSalaryGrades
      parameters:
        - name: organizationType
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SalaryGrade" }
    post:
      tags: [payroll]
      summary: Create salary grade
      operationId: createSalaryGrade
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SalaryGradeInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalaryGrade" }
  /salary-grades/{id}:
    get:
      tags: [payroll]
      summary: Get salary grade
      operationId: getSalaryGrade
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalaryGrade" }
    patch:
      tags: [payroll]
      summary: Update salary grade
      operationId: updateSalaryGrade
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SalaryGradeInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalaryGrade" }
    delete:
      tags: [payroll]
      summary: Delete salary grade
      operationId: deleteSalaryGrade
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: No Content

  # ── Pay Components ────────────────────────────────────────────────────────
...
  /payroll-runs:
    get:
      tags: [payroll]
      summary: List payroll runs
      operationId: listPayrollRuns
      parameters:
        - name: periodId
          in: query
          schema: { type: number }
        - name: employeeId
          in: query
          schema: { type: number }
        - name: status
          in: query
          schema: { type: string }
        - name: hasException
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PayrollRunSummary" }
  /payroll-runs/{id}:
    get:
      tags: [payroll]
      summary: Get payroll run detail with line items
      operationId: getPayrollRun
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayrollRunDetail" }
    patch:
      tags: [payroll]
      summary: Update exception note on a payroll run
      operationId: updatePayrollRun
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PayrollRunPatch" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayrollRunDetail" }
  /payroll-runs/{id}/payslip:
    get:
      tags: [payroll]
      summary: Get payslip view for a run
      operationId: getPayslip
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payslip" }
components:
  schemas:
...
    PublicHoliday:
      type: object
      properties:
        id: { type: number }
        nameEn: { type: string }
        nameAr: { type: string }
        date: { type: string }
        year: { type: number }
        isRecurring: { type: boolean }
        applicableTo: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, nameEn, nameAr, date, year, isRecurring, applicableTo, createdAt]
    PublicHolidayInput:
      type: object
      properties:
        nameEn: { type: string }
        nameAr: { type: string }
        date: { type: string }
        year: { type: number }
        isRecurring: { type: boolean }
        applicableTo: { type: string }
        notes: { type: ["string", "null"] }
      required: [nameEn, nameAr, date, year]

    # ── Salary Grades ─────────────────────────────────────────
    SalaryGrade:
      type: object
      properties:
        id: { type: number }
        gradeCode: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        step: { type: number }
        baseSalary: { type: string }
        housingAllowancePct: { type: string }
        transportAllowancePct: { type: string }
        currency: { type: string }
        organizationType: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }
      required: [id, gradeCode, nameEn, nameAr, step, baseSalary, housingAllowancePct, transportAllowancePct, currency, organizationType, isActive, createdAt]
    SalaryGradeInput:
      type: object
      properties:
        gradeCode: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        step: { type: number }
        baseSalary: { type: number }
        housingAllowancePct: { type: number }
        transportAllowancePct: { type: number }
        currency: { type: string }
        organizationType: { type: string }
        isActive: { type: boolean }
      required: [gradeCode, nameEn, nameAr, baseSalary]

    # ── Pay Components ────────────────────────────────────────
...
    PayrollRunLine:
      type: object
      properties:
        id: { type: number }
        payrollRunId: { type: number }
        payComponentId: { type: ["number", "null"] }
        codeEn: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        type: { type: string }
        amount: { type: string }
        sortOrder: { type: number }
      required: [id, payrollRunId, codeEn, nameEn, nameAr, type, amount, sortOrder]
    PayrollRunSummary:
      type: object
      properties:
        id: { type: number }
        payrollPeriodId: { type: number }
        employeeId: { type: number }
        employeeNameEn: { type: string }
        employeeNameAr: { type: string }
        employeeNumber: { type: string }
        jobTitleEn: { type: string }
        departmentNameEn: { type: string }
        grade: { type: ["string", "null"] }
        baseSalary: { type: string }
        grossSalary: { type: string }
        totalDeductions: { type: string }
        netSalary: { type: string }
        overtimeHours: { type: string }
        overtimePay: { type: string }
        currency: { type: string }
        hasException: { type: boolean }
        exceptionNote: { type: ["string", "null"] }
        status: { type: string }
        calculatedAt: { type: ["string", "null"] }
      required: [id, payrollPeriodId, employeeId, employeeNameEn, baseSalary, grossSalary, totalDeductions, netSalary, currency, hasException, status]
    PayrollRunDetail:
      allOf:
        - $ref: "#/components/schemas/PayrollRunSummary"
        - type: object
          properties:
            jobTitleAr: { type: string }
            nationalId: { type: string }
            periodNameEn: { type: string }
            periodNameAr: { type: string }
            periodStartDate: { type: string }
            periodEndDate: { type: string }
            payDate: { type: string }
            workingDays: { type: number }
            presentDays: { type: number }
            absentDays: { type: number }
            deductedLeaveDays: { type: string }
            leaveDeductionAmount: { type: string }
            totalEarnings: { type: string }
            lines:
              type: array
              items: { $ref: "#/components/schemas/PayrollRunLine" }
    PayrollRunPatch:
      type: object
      properties:
        exceptionNote: { type: ["string", "null"] }
        hasException: { type: boolean }
        status: { type: string }
    Payslip:
      type: object
      properties:
        runId: { type: number }
        employee:
          type: object
          properties:
            id: { type: number }
            employeeNumber: { type: string }
            fullNameEn: { type: string }
            fullNameAr: { type: string }
            jobTitleEn: { type: string }
            jobTitleAr: { type: string }
            nationalId: { type: string }
            grade: { type: ["string", "null"] }
            departmentNameEn: { type: string }
            departmentNameAr: { type: string }
          required: [employeeNumber, fullNameEn, fullNameAr, jobTitleEn, jobTitleAr, nationalId, departmentNameEn, departmentNameAr]
        period:
          type: object
          properties:
            periodCode: { type: string }
            nameEn: { type: string }
            nameAr: { type: string }
            startDate: { type: string }
            endDate: { type: string }
            payDate: { type: string }
          required: [periodCode, nameEn, nameAr, startDate, endDate, payDate]
        summary:
          type: object
          properties:
            baseSalary: { type: string }
            grossSalary: { type: string }
            totalEarnings: { type: string }
            totalDeductions: { type: string }
            netSalary: { type: string }
            overtimeHours: { type: string }
            overtimePay: { type: string }
            workingDays: { type: number }
            presentDays: { type: number }
            currency: { type: string }
          required: [baseSalary, grossSalary, totalEarnings, totalDeductions, netSalary, overtimeHours, overtimePay, workingDays, presentDays, currency]
        earnings:
          type: array
          items: { $ref: "#/components/schemas/PayrollRunLine" }
        deductions:
          type: array
          items: { $ref: "#/components/schemas/PayrollRunLine" }
        hasException: { type: boolean }
        exceptionNote: { type: ["string", "null"] }
        calculatedAt: { type: ["string", "null"] }
      required: [runId, employee, period, summary, earnings, deductions, hasException]