# lib/api-zod/src/generated/types/payrollRunSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface PayrollRunSummary {
  id: number;
  payrollPeriodId: number;
  employeeId: number;
  employeeNameEn: string;
  employeeNameAr?: string;
  employeeNumber?: string;
  jobTitleEn?: string;
  departmentNameEn?: string;
  /** @nullable */
  grade?: string | null;
  baseSalary: string;
  grossSalary: string;
  totalDeductions: string;
  netSalary: string;
  overtimeHours?: string;
  overtimePay?: string;
  currency: string;
  hasException: boolean;
  /** @nullable */
  exceptionNote?: string | null;
  status: string;
  /** @nullable */
  calculatedAt?: string | null;
}