# lib/api-zod/src/generated/types/payslip.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise HRMS API
 * OpenAPI spec version: 0.1.0
 */
import type { PayrollRunLine } from './payrollRunLine';
import type { PayslipEmployee } from './payslipEmployee';
import type { PayslipPeriod } from './payslipPeriod';
import type { PayslipSummary } from './payslipSummary';

export interface Payslip {
  runId: number;
  employee: PayslipEmployee;
  period: PayslipPeriod;
  summary: PayslipSummary;
  earnings: PayrollRunLine[];
  deductions: PayrollRunLine[];
  hasException: boolean;
  /** @nullable */
  exceptionNote?: string | null;
  /** @nullable */
  calculatedAt?: string | null;
}