# lib/api-zod/src/generated/types/salaryGrade.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface SalaryGrade {
  id: number;
  gradeCode: string;
  nameEn: string;
  nameAr: string;
  step: number;
  baseSalary: string;
  housingAllowancePct: string;
  transportAllowancePct: string;
  currency: string;
  organizationType: string;
  isActive: boolean;
  createdAt: string;
}